PrintMax
The stored procedure created by this script prints a text of any size, even if the size is greater than 8,000 bytes, which is the limit of the regular PRINT statement. It splits the large text into chunks of up to 4,000 characters, and then uses the regular PRINT statement. It searches for line breaks in the text, and prefers the split to occur where there are line breaks, so that the output is more readbale.